refactor(crawler): move run_crawl_task + handle_crawl_result to crawl_task.rs (strangler fig B)
